+++ scripts/train.py
@@ -5,7 +5,8 @@ from datetime import datetime
 import wandb
 from datasets import concatenate_datasets
 
-from arcsf.config.experiment import EXPERIMENT_CONFIG_DIR, ExperimentConfig
+from arcsf.config.experiment import ExperimentConfig
+from arcsf.constants import EXPERIMENT_CONFIG_DIR
 from arcsf.data.tofu import load_tofu
 from arcsf.models.model import load_model_and_tokenizer
 from arcsf.models.trainer import load_trainer
